backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
python-multipart==0.0.12
pyyaml==6.0.2
python-dateutil==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
